newswordy/requirements.txt
django
newsapi-python
python-dateutil
requests